+++ Interface/AddOns/PanzaUI/options.lua
@@ -116,7 +116,7 @@ addCheckbox("Configure Raid & Party", "Resize and colour Role icons.", "configRa
 yOffset = yOffset - 28
 addCheckbox("Configure Target Frame", "Hide background and remove buffs/debuffs on target.", "configTargetFrame", yOffset)
 yOffset = yOffset - 28
-addCheckbox("Configure Various frames", "Reskin Tooltip's Healthbar, etc.", "configVariousFrames", yOffset)
+addCheckbox("Configure Various frames", "Reskin Tooltip and Nameplate's healthbar, etc.", "configVariousFrames", yOffset)
 yOffset = yOffset - 28
 
 createSectionHeader("Hide Frames", yOffset)