+++ CHANGELOG.md
@@ -17,7 +17,8 @@
 * Moved Minimap text aligning it to the Calendar icon
 * Optimized the code (now using less than 20KB of memory)
 * Removed Queue Icon configuration option
-* Reskin Tooltip's Healthbar
+* Reskin Nameplate's healthbar
+* Reskin Tooltip's healthbar
 * Revamped Add-on window in-game to better show the various options
 
 ### ⌨️ Macros